back/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
PyJWT==2.8.0
stripe==7.8.0
django-storages==1.14.2
boto3==1.34.14
whitenoise==6.6.0
gunicorn==21.2.0
daphne==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
scikit-learn==1.4.0
google-auth==2.26.1
django-extensions==3.2.3